+++ packages/react/src/atomic/Box/src/box.stories.tsx
@@ -1,5 +1,7 @@
 import React, { forwardRef } from "react";
 import type { Meta, StoryObj } from "@storybook/react";
+import { box } from "@nimbus-ds/styles";
+import { argTypesConvert } from ".storybook/utils";
 import { Box as BoxComponent, BoxProps } from "./Box";
 
 export const Basic: React.FC<BoxProps> = forwardRef((props: BoxProps) => (
@@ -20,6 +22,7 @@ const meta: Meta<typeof Basic> = {
         "A ref to the element rendered by this component. Because this component is polymorphic, the type will vary based on the value of the as prop.",
     },
     children: { control: { type: "text" } },
+    ...argTypesConvert(box.properties),
   },
   tags: ["autodocs"],
 };