+++ packages/react/src/composite/Card/src/components/CardFooter/CardFooter.tsx
@@ -6,10 +6,17 @@ import { CardFooterProps } from "./cardFooter.types";
 const CardFooter: React.FC<CardFooterProps> = ({
   className: _className,
   style: _style,
+  padding = "none",
   children,
   ...rest
 }) => (
-  <div className={card.classnames.container__footer} {...rest}>
+  <div
+    className={[
+      card.classnames.container__footer,
+      card.subComponents.footer.sprinkle({ padding }),
+    ].join(" ")}
+    {...rest}
+  >
     {children}
   </div>
 );