+++ packages/react/src/composite/Card/src/components/CardHeader/CardHeader.tsx
@@ -1,16 +1,18 @@
 import React from "react";
 import { Title } from "@nimbus-ds/title";
+import { card } from "@nimbus-ds/styles";
 
 import { CardHeaderProps } from "./cardHeader.types";
 
 const CardHeader: React.FC<CardHeaderProps> = ({
   className: _className,
   style: _style,
+  padding = "none",
   title,
   children,
   ...rest
 }) => (
-  <div {...rest}>
+  <div className={card.subComponents.header.sprinkle({ padding })} {...rest}>
     {title && (
       <Title data-testid="header-title" as="h3">
         {title}