+++ packages/react/src/composite/Modal/src/components/ModalFooter/ModalFooter.tsx
@@ -6,10 +6,17 @@ import { ModalFooterProps } from "./modalFooter.types";
 const ModalFooter: React.FC<ModalFooterProps> = ({
   className: _className,
   style: _style,
+  padding = "none",
   children,
   ...rest
 }) => (
-  <div className={modal.classnames.container__footer} {...rest}>
+  <div
+    className={[
+      modal.subComponents.body.sprinkle({ padding }),
+      modal.classnames.container__footer,
+    ].join(" ")}
+    {...rest}
+  >
     {children}
   </div>
 );