+++ packages/react/src/composite/Modal/src/components/ModalFooter/modalFooter.types.ts
@@ -1,4 +1,5 @@
 import { HTMLAttributes, ReactNode } from "react";
+import { modal } from "@nimbus-ds/styles";
 
 export interface ModalFooterProperties {
   /**
@@ -6,6 +7,11 @@ export interface ModalFooterProperties {
    * @TJS-type React.ReactNode
    */
   children: ReactNode;
+  /**
+   * The padding properties are used to generate space around an modal's footer content area.
+   * @default none
+   */
+  padding?: keyof typeof modal.subComponents.footer.properties.padding;
 }
 
 export type ModalFooterProps = ModalFooterProperties &