+++ packages/react/src/composite/Modal/src/components/ModalHeader/modalHeader.types.ts
@@ -1,4 +1,5 @@
 import { HTMLAttributes, ReactNode } from "react";
+import { modal } from "@nimbus-ds/styles";
 
 export interface ModalHeaderProperties {
   /**
@@ -10,6 +11,11 @@ export interface ModalHeaderProperties {
    * The title to display in the modal header.
    */
   title?: string;
+  /**
+   * The padding properties are used to generate space around an modal's header content area.
+   * @default none
+   */
+  padding?: keyof typeof modal.subComponents.header.properties.padding;
 }
 
 export type ModalHeaderProps = ModalHeaderProperties &